Code quality fixes: bug fixes, deprecated API updates, and cleanup
- Fix dead return statement in core/utils.py (build_monitor_result_details)
- Fix garbled Korean text in main.py closeEvent log message
- Replace deprecated numpy.float with float in json_checker_new.py
- Replace deprecated ssl.wrap_socket() with SSLContext in server_thread.py
- Unify result_dir to use CONSTANTS.result_dir in platformVal_all/systemVal_all
- Add threading.Lock for rows.json concurrent access in server_thread/platformVal_all
- Remove duplicate WebhookBadgeLabel class from common_main_ui.py (import from gui_utils)
- Replace all print() calls with Logger in validation_registry.py
- Remove duplicate importlib import in validation_registry.py
- Remove deprecated cgi module, replace cgi.parse_header with get_content_type()
- Remove unused import logging in main.py
- Remove duplicate Logger import in common_main_ui.py
- Delete unrelated scripts from ui/pages (ACCV paper analysis, screen automation)

+++ core/json_checker_new.py
@@ -704,7 +704,7 @@ def do_checker(all_field, datas, opt_field, flag_opt):
                             else:
                                 if field[-1] == int:
                                     if type(raw_data[-1]) == numpy.int64 or type(raw_data[-1]) == numpy.int32 or type(
-                                            raw_data[-1]) == numpy.float:
+                                            raw_data[-1]) == float:
                                         raw_data[-1] = True
                                     else:
                                         raw_data[-1] = "Value Type Error: " + str(field[1]) + " " + str(
